% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_role.R
\name{add_role}
\alias{add_role}
\title{Add a role to the database}
\usage{
add_role(conn, username, group)
}
\arguments{
\item{conn}{Connection to a database using admin credentials. Currently this will work with only the lnrs and shared databases.}

\item{username}{The username for the new credentials.}

\item{group}{The name of a user group to add the credentials to.}
}
\value{
A list including the username and password of the new credentials
}
\description{
Creates a new role and records it in an admin.logins table.
}
