% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCustomLegend.R
\name{addCustomLegend}
\alias{addCustomLegend}
\title{Title
Create a custom legend for a leaflet webmap}
\usage{
addCustomLegend(map, colors, labels, sizes, shapes, borders, opacity = 0.5)
}
\arguments{
\item{map}{the leaflet map to add the legend to}

\item{colors}{the colors, ideally as hex codes}

\item{labels}{the character labels}

\item{sizes}{the sizes, as a numeric vector}

\item{shapes}{the character shapes}

\item{borders}{the border colours, ideally as hex codes}

\item{opacity}{the opacity, ranging from 0 (transparent) to 1 (opaque)}
}
\description{
Title
Create a custom legend for a leaflet webmap
}
\examples{
colors <- c("#E69F00", "#56B4E9", "#009E73")
labels <- c("One", "Two", "Three")
sizes <- c(10, 5, 7.5)
shapes <- c("square", "circle", "square")
borders <- c("#E69F00", "#56B4E9", "#009E73")
leaflet::leaflet() \%>\%
  addCustomLegend(colors, labels, sizes, shapes, borders)
}
