% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{create_db}
\usage{
create_db(conn, db_name, app_role)
}
\arguments{
\item{conn}{A DBI database connection connecting to the postgres database with the admin login}

\item{db_name}{The name of the database to be created}

\item{app_role}{The name of the user to be created with the database}
}
\value{
The details of the database and user created
}
\description{
create_db
}
\examples{
conn <- pool::dbPool(drv = RPostgres::Postgres(), host = "xxxxxxx", port = 5432, dbname = "postgres", user = "xxxxx", password = "xxxxxxx", sslmode = "prefer")
create_db(conn, "test_db", "test_user")
}
