% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_data_model.R
\name{db_to_json_data_model}
\alias{db_to_json_data_model}
\title{Create a JSON data model based on a database}
\usage{
db_to_json_data_model(conn, file_path)
}
\arguments{
\item{conn}{database connection object}

\item{file_path}{path to save the JSON file}
}
\value{
a JSON file with the data model
}
\description{
This function helps with viewing database structures, by providing the right json to be ready by \code{read_json_data_model}.
You can then make destructive updates to the data structure using \code{\link{db_write_data_model}}.
}
\examples{
\dontrun{
 conn <- DBI::dbConnect(odbc::odbc(), dsn = "Legacy Landscapes", timeout = 10)
 db_to_json_data_model(conn, "data_model.json")
 m <- read_json_data_model("data_model.json")
 m <- model_relational_data(m)
 db_write_data_model(m, schema_name = NULL, crs_srid = 4326, overwrite = T, conn = conn)
 DBI::dbDisconnect(conn)
 }


}
