% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_data_model.R
\name{db_write_data_model}
\alias{db_write_data_model}
\title{Write a ReDaMoR data model to the database}
\usage{
db_write_data_model(x, schema_name, owner_name, overwrite = F, conn = db)
}
\arguments{
\item{x}{data model generated using ReDaMoR::read_json_data_model or ReDaMoR::model_relational_data}

\item{schema_name}{name of the scheme to create the tables in}

\item{owner_name}{owner of the tables}

\item{overwrite}{logical. If set to TRUE then it will overwrite any existing tables in the schema with the same name as a table in the model.}

\item{conn}{database connection}
}
\description{
Write a ReDaMoR data model to the database
}
