% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en2os.R
\name{en2os}
\alias{en2os}
\title{Eastings and Northings to OS Grid Ref
This function converts eastings and northings to an OS grid ref of a specified length.}
\usage{
en2os(easting, northing, resolution)
}
\arguments{
\item{easting}{The eastings of the point to be converted.}

\item{northing}{The northings of the point to be converted.}

\item{resolution}{The resolution of the grid reference to be produced.}
}
\description{
Eastings and Northings to OS Grid Ref
This function converts eastings and northings to an OS grid ref of a specified length.
}
\examples{
df <- data.frame(easting = c(385760, 532765, 479732), northing = c(219790, 122444, 347045))
gridrefs <- en2os(df$easting, df$northing, 6)
}
\keyword{Grid}
\keyword{references}
