% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{execute_lazy}
\alias{execute_lazy}
\title{Run a query and return the result as a data frame or sf}
\usage{
execute_lazy(lazy_tbl, limit = 1000)
}
\arguments{
\item{x}{a dplyr tbl object or a character string containing a SQL query.}

\item{ref}{a list containing the table schema and name.}

\item{conn}{a connection object. Only required if x is a character string.}
}
\value{
a data frame or sf object
}
\description{
Run a query and return the result as a data frame or sf
}
