% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definition.R
\name{get_definition}
\alias{get_definition}
\title{Get the Definition of a View or Materialised View}
\usage{
get_definition(ref, conn = db)
}
\arguments{
\item{ref}{a table reference in the format provided by list_tables()}

\item{conn}{a connection object}
}
\value{
a character string containing the definition of the view or materialised view in SQL
}
\description{
Get the Definition of a View or Materialised View
}
