% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lazy.R
\name{get_lazy}
\alias{get_lazy}
\title{Get a lazy data frame for a given table or view reference.}
\usage{
get_lazy(ref, conn = db$conn)
}
\arguments{
\item{ref}{a table reference from db_connect}

\item{conn}{a connection object from db_connect. Defaults to \code{db$conn}.}
}
\value{
a dplyr tbl object
}
\description{
Get a lazy data frame for a given table or view reference.
}
\details{
This function generates a dplyr lazy tbl object from a table or view reference in a listed_connection object generated by db_connect.
This is useful for creating an sql query with dplyr syntax.
}
