% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwt_db_connect.R
\name{kwt_db_connect}
\alias{kwt_db_connect}
\title{kwt_db_connect}
\usage{
kwt_db_connect(
  db = "shared",
  username = Sys.getenv("user"),
  password = Sys.getenv("password")
)
}
\arguments{
\item{db}{name of the internal database within postgres, shared is the standard}

\item{username}{postgres username, default autoreads kwt_user from global .renviron file, see usethis::edit_r_environ() for details}

\item{password}{postgres password, default autoreads kwt_password from global .renviron file, see usethis::edit_r_environ() for details}
}
\description{
kwt_db_connect
}
\examples{
conn <- kwt_db_connect()
}
