% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_db.R
\name{list_db}
\alias{list_db}
\title{List all schemas, tables and views in a database}
\usage{
list_db(conn = db)
}
\arguments{
\item{conn}{a connection object}
}
\value{
a list of schemas. Each schema is a list of refs for tables, views and materialised views in that schema. Each ref is a list containing three values:
\itemize{
\item table_schema: the schema the table/view is in
\item table_name: the name of the table/view
\item type: the type of the table/view. Either "table", "view" or "matview"
}
}
\description{
List all schemas, tables and views in a database
}
