% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_views.R
\name{map_views}
\alias{map_views}
\title{View a directed graph of view dependencies in a database.}
\usage{
map_views(
  conn = db,
  ignore_schemas = c("information_schema", "pg_catalog", "postgis", "admin", "dev",
    "outputs_backup", "public")
)
}
\arguments{
\item{conn}{A DBI connection object.}

\item{ignore_schemas}{A character vector of schema names to ignore.}
}
\value{
An htmltools widget showing the directed graph of view dependencies.
}
\description{
View a directed graph of view dependencies in a database.
}
