% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os2en.R
\name{os2en}
\alias{os2en}
\title{OS Grid Ref to Eastings and Northings}
\usage{
os2en(gridreference)
}
\arguments{
\item{gridreference}{The grid reference to be converted. Can be any resolution.}
}
\description{
This function creates a data frame containing the easting and northing coordinates for the SW corner of the grid reference.
}
\examples{
df <- data.frame(gridref = c("SO857197", "TQ327224", "SK797470"))
coords <- os2en(df$gridref)
}
\keyword{Grid}
\keyword{references}
