% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os2point.R
\name{os2point}
\alias{os2point}
\title{Function for creating a polygon from a grid reference, with the size of the polygon dependant on the precision of the grid reference}
\usage{
os2point(df, grid_ref_col = "grid_ref")
}
\arguments{
\item{df}{The data frame, must include a column containing grid references}

\item{grid_ref_col}{The name of the column containing the grid references}
}
\description{
This function creates an sf polygon object, .
}
\examples{
df <- data.frame(gridref = c("SO857197", "TQ327224", "SK797470"))
coords <- os2point(df, grid_ref_col = "gridref")
}
