% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os2poly.R
\name{os2poly}
\alias{os2poly}
\title{Function for creating a polygon from a grid reference, with the size of the polygon dependent on the precision of the grid reference}
\usage{
os2poly(df, grid_ref_col = "grid_ref")
}
\arguments{
\item{df}{The data frame, must include a column containing grid references}

\item{grid_ref_col}{The name of the column containing the grid references}
}
\description{
This function creates an sf polygon object, .
}
\examples{
df <- data.frame(gridref = c("SO85719789", "TQ327224", "SK7974", "TF73G"))
polys <- os2poly(df, grid_ref_col = "gridref")
}
