% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode2point.R
\name{postcode2point}
\alias{postcode2point}
\title{Convert postcode to sf point}
\usage{
postcode2point(df, postcode_col)
}
\arguments{
\item{df}{input data frame, containing postcode values}

\item{postcode_col}{the name of the column containing the postcode}
}
\value{
an sf data frame, containing a formatted postcodes column alongside the original
}
\description{
Convert postcode to sf point
}
\examples{
postcodes_df <- data.frame(postcode = c("AB101AB", "AB10 1AF", "ab10 1ag"))
postcodes_sf <- postcode2point(postcodes_df, "postcode")
}
