% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref_lookup.R, R/postcode_lookup.R
\docType{data}
\name{postcode_lookup}
\alias{postcode_lookup}
\title{Lookup table with easting and northing coordinates for each UK postcode.}
\format{
A data frame with 1725632 rows and 3 variables:
\describe{
\item{postcode}{uk postcode, with no spaces and capitalised}
\item{eastings}{easting coordinate}
\item{northings}{northing coordinate}
...
}

A data frame with 1725632 rows and 3 variables:
\describe{
\item{postcode}{uk postcode, with no spaces and capitalised}
\item{eastings}{easting coordinate}
\item{northings}{northing coordinate}
...
}
}
\source{
\url{https://osdatahub.os.uk/downloads/open/CodePointOpen}

\url{https://osdatahub.os.uk/downloads/open/CodePointOpen}
}
\usage{
postcode_lookup

postcode_lookup
}
\description{
A dataset containing the easting and northing coordinates of UK postcodes.

A dataset containing the easting and northing coordinates of UK postcodes.
}
\keyword{datasets}
