% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_query.R
\name{start_query}
\alias{start_query}
\title{Create a query from a table reference}
\usage{
start_query(ref, conn = db)
}
\arguments{
\item{ref}{a table reference in the format provided by list_tables()}

\item{conn}{a connection object}
}
\value{
a dplyr tbl object
}
\description{
Create a query from a table reference
}
\details{
This function generates a dplyr lazy tbl object from a table reference in the format provided by list_tables().
This is useful for creating an sql query with dplyr syntax.
}
