% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{`\%nin\%`}
\alias{`\%nin\%`}
\alias{\%nin\%}
\title{Negation of \code{\%in\%}}
\usage{
... \%nin\% NA
}
\value{
A logical vector indicating if a match was not found.
}
\description{
This function returns the negation of the \code{\%in\%} operator.
}
\examples{
1 \%nin\% c(2, 3, 4)
"a" \%nin\% c("b", "c", "d")
}
