% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_title_case.R
\name{to_title_case}
\alias{to_title_case}
\title{Convert a string from snake_case to a pretty title case}
\usage{
to_title_case(name)
}
\arguments{
\item{name}{A string in snake_case}
}
\value{
A character vector of names in title case
}
\description{
Takes a character vector and converts it to a character vector of nice looking titles,
taking into account common acronyms and abbreviations.
}
\examples{
to_title_case("my_column_name")
# "My Column Name"

to_title_case(c("kwtid", "img_url", "geom"))
#  "KWTID"     "Image URL" "Geometry"

}
