% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_report.R
\name{write_table_report}
\alias{write_table_report}
\title{Write an HTML report about a table}
\usage{
write_table_report(
  ref,
  conn = db,
  filepath = paste0(ref$table_schema, "__", ref$table_name, ".html")
)
}
\arguments{
\item{ref}{A reference to a table}

\item{conn}{A database connection}

\item{filepath}{Path to an html file to write the report as. Default is \code{table_schema__table_name.html}}
}
\description{
Write an HTML report about a table
}
