% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_data_model.R
\name{db_to_json_data_model}
\alias{db_to_json_data_model}
\title{Create a JSON data model based on a database}
\usage{
db_to_json_data_model(x, file_path)
}
\arguments{
\item{x}{database connection object}

\item{file_path}{path to save the JSON file}
}
\value{
a JSON file with the data model
}
\description{
Create a JSON data model based on a database
}
