% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_vals.R
\name{edit_vals}
\alias{edit_vals}
\title{Edit values of a table.}
\usage{
edit_vals(ref, backup = T, conn = db)
}
\arguments{
\item{ref}{a table reference in the format provided by list_tables()}

\item{backup}{whether to make a backup of the data before editing}

\item{conn}{a connection object}
}
\description{
Edit values of a table.
}
