% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_json_data_model}
\alias{model_relational_data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{postgisMoR}{\code{\link[postgisMoR]{model_relational_data}}, \code{\link[postgisMoR]{read_json_data_model}}}
}}

